#1 - pwm with timer1
