+++ main/my_callback.h
@@ -1,0 +1,10 @@
+#ifndef MY_CALLBACK_FUNCTION
+
+#define MY_CALLBACK_FUNCTION
+
+#include "my_lib.h"
+
+void timer_cb(void* arg);
+
+#endif
+
