+++ main/my_lib.h
@@ -1,0 +1,21 @@
+#ifndef MY_LIBRARY
+
+#define MY_LIBRARY
+
+#include <stdbool.h>
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+
+#include "driver/gpio.h"
+#include "driver/hw_timer.h"
+
+#include "pwm_timer1.h"
+#include "my_callback.h"
+
+#include "esp_now.h"
+#include "esp_log.h"
+#include "esp_wifi.h"
+
+
+#endif
+
